% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{aboutUI}
\alias{aboutUI}
\title{About module UI}
\usage{
aboutUI(id)
}
\arguments{
\item{id}{A string identifier for the module namespace.}
}
\value{
A Shiny UI element (\code{fluidPage}) for rendering the about page.
}
\description{
Provides a static informational page describing the purpose and functionality
of the B1MG Variant Voting application. This module does not require server-side logic.
}
