% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_admin.R
\name{adminServer}
\alias{adminServer}
\title{Admin module server}
\usage{
adminServer(id, cfg, login_trigger, db_pool, tab_trigger = NULL)
}
\arguments{
\item{id}{Module namespace}

\item{cfg}{App configuration}

\item{login_trigger}{Reactive containing login data}

\item{db_pool}{Database connection pool}

\item{tab_trigger}{Optional reactive triggered when admin tab is selected}
}
\description{
Shows password retrieval tokens for users, allows admins to add new users,
enables downloading user annotations, and provides functionality to reset user annotations.
}
