% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{cancel_pending_logout}
\alias{cancel_pending_logout}
\title{Cancel a pending logout update}
\usage{
cancel_pending_logout(sessionid)
}
\arguments{
\item{sessionid}{Character. The session ID to cancel logout for}
}
\description{
Cancel a pending logout update
}
