% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_env_utils.R
\name{copy_dir_from_app}
\alias{copy_dir_from_app}
\title{Copy a directory from the app to a target path
This function copies a directory from the app's default environment to a specified target path.}
\usage{
copy_dir_from_app(target_dir_path)
}
\arguments{
\item{target_dir_path}{Character. Path to the target directory where the app directory will be copied.}
}
\value{
Character path to the copied directory
}
\description{
Copy a directory from the app to a target path
This function copies a directory from the app's default environment to a specified target path.
}
\keyword{internal}
