% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_database}
\alias{create_database}
\title{Initialize external database}
\usage{
create_database(db_path, to_be_voted_images_file, grouped_credentials_file)
}
\arguments{
\item{base_dir}{Character. Base directory for database (default: current working directory)}

\item{db_name}{Character. Name of the database file (default: "db.sqlite")}
}
\value{
Character path to the database file
}
\description{
Sets up the SQLite database file outside the package using the same logic as init_db.R.
}
