% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{ensure_gitignore}
\alias{ensure_gitignore}
\title{This function checks if a .gitignore file exists in the specified directory.
If it does not exist, it creates one. It then ensures that the specified patterns
are present in the .gitignore file, adding them if they are missing.}
\usage{
ensure_gitignore(dir, patterns)
}
\arguments{
\item{dir}{Character. Directory path where the .gitignore file should be checked/created.}

\item{patterns}{Character vector. Patterns to ensure in the .gitignore file.}
}
\value{
Character path to the .gitignore file
}
\description{
This function checks if a .gitignore file exists in the specified directory.
If it does not exist, it creates one. It then ensures that the specified patterns
are present in the .gitignore file, adding them if they are missing.
}
\keyword{internal}
