% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_faq.R
\name{faqServer}
\alias{faqServer}
\title{FAQ module server logic}
\usage{
faqServer(id, cfg, tab_trigger = NULL)
}
\arguments{
\item{id}{A string identifier for the module namespace.}

\item{cfg}{App configuration}

\item{tab_trigger}{Optional reactive that triggers when the about tab is selected}
}
\value{
None. Side effect only: registers a module server.
}
\description{
FAQ module server logic
}
