% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_faq.R
\name{faqUI}
\alias{faqUI}
\title{FAQ module UI}
\usage{
faqUI(id, cfg)
}
\arguments{
\item{id}{A string identifier for the module namespace.}
}
\value{
A Shiny UI element (\code{fluidPage}) for rendering the about page.
}
\description{
Provides a static informational page describing the purpose and functionality
of the B1MG Variant Voting application.
}
