% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file.symlink.exists}
\alias{file.symlink.exists}
\title{Check if a file is a symlink}
\usage{
file.symlink.exists(path)
}
\arguments{
\item{path}{Character. Path to check}
}
\value{
Logical. TRUE if path is a symlink, FALSE otherwise
}
\description{
Check if a file is a symlink
}
