% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_env_utils.R
\name{generate_password}
\alias{generate_password}
\title{Generate a random password}
\usage{
generate_password(length = 12, pattern = "!@#$\%^&*")
}
\arguments{
\item{length}{Integer. Length of the password to generate. Defaults to 12.}

\item{pattern}{Character string. characters to include in the password.}
}
\value{
Character string containing the generated password
}
\description{
Creates a random password of specified length using letters, numbers, and special characters.
}
