% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_env_utils.R
\name{generate_password_retrieval_link}
\alias{generate_password_retrieval_link}
\title{Generate a password retrieval link}
\usage{
generate_password_retrieval_link(userid)
}
\arguments{
\item{token}{Character. Password retrieval token.}
}
\value{
Character string of the form \verb{?token=<token>}
}
\description{
Creates a query string containing the provided token. The resulting string
can be appended to a URL so that the token is available as a query
parameter.
}
