% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{get_app_dir}
\alias{get_app_dir}
\title{Get the path to the Shiny app directory}
\usage{
get_app_dir()
}
\value{
Character path to the Shiny app directory
}
\description{
Get the path to the Shiny app directory
}
