% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{init_db}
\alias{init_db}
\title{Initialize the database connection pool}
\usage{
init_db(cfg_sqlite_file)
}
\arguments{
\item{cfg_sqlite_file}{Path to an existing SQLite file}
}
\value{
A DBI pool object
}
\description{
Initialize the database connection pool
}
\keyword{internal}
