% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{init_environment}
\alias{init_environment}
\title{Initialize environment for the application}
\usage{
init_environment(config_file_path, base_dir = getwd())
}
\arguments{
\item{config_file_path}{Character. Path to the configuration file.
Default is app_env/config/config.yaml in the current working directory.}
}
\value{
List with paths to user_data directory, database file, and config file
}
\description{
Sets up the app environment directory and the database.
}
