% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_external_server_data}
\alias{init_external_server_data}
\title{Initialize external server_data directory}
\usage{
init_external_server_data(base_dir = getwd())
}
\arguments{
\item{base_dir}{Character. Base directory where server_data should be created}
}
\value{
Character. Path to the created server_data directory
}
\description{
Creates and sets up the external server_data directory structure.
}
