% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_leaderboard.R
\name{leaderboardUI}
\alias{leaderboardUI}
\title{Leaderboard module UI}
\usage{
leaderboardUI(id, cfg)
}
\arguments{
\item{id}{A string identifier for the module namespace.}
}
\value{
A Shiny UI element (\code{fluidPage}) for displaying the leaderboard.
}
\description{
Provides a user interface for displaying the number of votes submitted
by each participating institute in the B1MG Variant Voting app.
}
\details{
The UI includes a table of vote counts and a button to refresh the data.
}
