% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config_bk}
\alias{load_config_bk}
\title{Load voting app configuration}
\usage{
load_config_bk(config_dir)
}
\value{
A named list of configuration values (e.g., \code{cfg_sqlite_file}, \code{cfg_radio_options2val_map}, etc.)
}
\description{
Returns a list of configuration values, optionally overridden by environment
variables or external files.
}
