% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login.R
\name{loginUI}
\alias{loginUI}
\title{Login module UI}
\usage{
loginUI(id, cfg)
}
\arguments{
\item{id}{A string identifier for the module namespace.}
}
\value{
A Shiny UI element (typically a login panel) rendered within a namespace.
}
\description{
Provides the user interface for logging in to the B1MG Variant Voting application.
This module is based on \code{shinyauthr} and includes reactive user authentication.
}
