% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_server.R
\name{makeVotingAppServer}
\alias{makeVotingAppServer}
\title{Factory function to create the Shiny server for the voting application}
\usage{
makeVotingAppServer(db_pool, cfg)
}
\arguments{
\item{db_pool}{A database connection pool}
}
\value{
A Shiny server function
}
\description{
Factory function to create the Shiny server for the voting application
}
