% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_voting.R
\name{numberedRadioButtons}
\alias{numberedRadioButtons}
\title{Voting module UI}
\usage{
numberedRadioButtons(
  inputId,
  label,
  choices,
  selected = NULL,
  inline = FALSE,
  width = NULL,
  choiceNames = NULL,
  choiceValues = NULL
)
}
\arguments{
\item{id}{A string identifier for the module namespace.}
}
\value{
A Shiny UI element (\code{fluidPage}) representing the voting interface.
}
\description{
Provides the user interface for displaying a voting task, including:
\itemize{
\item An image of a candidate somatic mutation
\item A radio button to express agreement with the annotation
\item Conditional inputs for alternate variant type and comments
\item Navigation controls (Back / Next)
}
}
\details{
This module uses \code{shinyjs} for interactivity and includes a custom \code{hotkeys.js}
script to enable keyboard shortcuts (e.g., Enter for "Next", Backspace for "Back").

The displayed options and labels are configured using:
\itemize{
\item \code{cfg$radioBtns_label}
\item \code{cfg$radio_options2val_map}
\item \code{cfg$checkboxes_label}
\item \code{cfg$observations2val_map}
}

These should be defined in a sourced configuration file (config.yaml).
}
