% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout_utils.R
\docType{data}
\name{pending_logout_tasks}
\alias{pending_logout_tasks}
\title{Shared environment for logout scheduling tasks}
\format{
An object of class \code{environment} of length 0.
}
\usage{
pending_logout_tasks
}
\description{
Shared environment for logout scheduling tasks
}
\keyword{internal}
