% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{populate_annotations_table}
\alias{populate_annotations_table}
\title{Populate the annotations table with data from a file}
\usage{
populate_annotations_table(conn, to_be_voted_images_file)
}
\arguments{
\item{conn}{Database connection object}

\item{to_be_voted_images_file}{Character. Path to the file containing image annotations}
}
\description{
Populate the annotations table with data from a file
}
\keyword{internal}
