% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{populate_users_table}
\alias{populate_users_table}
\title{Populate the users table with data from the grouped credentials file}
\usage{
populate_users_table(conn, grouped_credentials)
}
\arguments{
\item{conn}{Database connection object}

\item{grouped_credentials}{Object. Loaded grouped credentials from the config file}
}
\description{
Populate the users table with data from the grouped credentials file
}
\keyword{internal}
