% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_voting_app}
\alias{run_voting_app}
\title{Run B1MG Variant Voting Shiny Application}
\usage{
run_voting_app(host = "127.0.0.1", port = NULL, launch.browser = TRUE, ...)
}
\arguments{
\item{host}{Character. Host to run the application on. Default is "127.0.0.1"}

\item{port}{Integer. Port to run the application on. Default is NULL (random port)}

\item{launch.browser}{Logical. Should the browser be launched? Default is TRUE}

\item{...}{Additional arguments passed to shiny::runApp()}
}
\value{
Runs the Shiny application
}
\description{
This function launches the B1MG Variant Voting Shiny application.
}
\examples{
\dontrun{
run_voting_app()
}
}
