% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_userstats.R
\name{userstatsUI}
\alias{userstatsUI}
\title{User statistics module UI}
\usage{
userstatsUI(id, cfg)
}
\arguments{
\item{id}{A string identifier for the module namespace.}
}
\value{
A Shiny UI element (\code{fluidPage}) for rendering user statistics.
}
\description{
Provides the user interface for displaying statistics related to the current user's
voting activity within the B1MG Variant Voting app. This typically includes a table
of vote counts or annotations per session, along with a refresh button.
}
