% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_admin.R
\name{adminUI}
\alias{adminUI}
\title{Admin module UI}
\usage{
adminUI(id, cfg)
}
\arguments{
\item{id}{Module namespace}

\item{cfg}{App configuration object}
}
\value{
A Shiny UI element
}
\description{
Displays password retrieval tokens for users who have not retrieved their password yet and
allows admins to add new users.
}
