% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{copy_dir_from_app}
\alias{copy_dir_from_app}
\title{Initialize external environment for the application}
\usage{
copy_dir_from_app(target_dir_path)
}
\arguments{
\item{base_dir}{Character. Base directory for external files.
Defaults to current working directory.}
}
\value{
List with paths to user_data directory, database file, and config file
}
\description{
Sets up the external directory structure, database, and configuration for the application.
}
