% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{create_database}
\alias{create_database}
\title{Initialize SQLite database}
\usage{
create_database(db_path, to_be_voted_images_file, grouped_credentials)
}
\arguments{
\item{db_path}{Character. Path to the SQLite database file to create}

\item{to_be_voted_images_file}{Character. Path to
the file containing image annotations}

\item{grouped_credentials}{List. Parsed YAML list of grouped credentials}
}
\value{
Character path to the database file
}
\description{
Initialize SQLite database
}
