% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_config.R
\name{load_config}
\alias{load_config}
\title{Load voting app configuration}
\usage{
load_config(config_file_path)
}
\arguments{
\item{config_file_path}{Character. Path to the configuration file. Default is the one in the app directory}
}
\value{
A named list of configuration values (e.g., \code{cfg_sqlite_file}, \code{cfg_radio_options2val_map}, etc.)
}
\description{
Returns a list of configuration values, optionally overridden by environment
variables or external files.
}
