% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votingAppServer.R
\name{makeVotingAppServer}
\alias{makeVotingAppServer}
\title{Make a Shiny app server function for the voting application}
\usage{
makeVotingAppServer(db_pool)
}
\arguments{
\item{db_pool}{A database connection pool}
}
\value{
A Shiny server function
}
\description{
Make a Shiny app server function for the voting application
}
\keyword{internal}
