% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{populate_users_table}
\alias{populate_users_table}
\title{Populate the users table with data from institute2userids2password.yaml
#' @keywords internal
#' @param conn Database connection object}
\usage{
populate_users_table(conn, grouped_credentials_file)
}
\arguments{
\item{grouped_credentials_file}{Character. Path to the grouped_credentials.yaml file}
}
\description{
Populate the users table with data from institute2userids2password.yaml
#' @keywords internal
#' @param conn Database connection object
}
